+++ src/types/stylish.d.ts
@@ -1,0 +1,3 @@
+declare module 'eslint/lib/cli-engine/formatters/stylish' {
+  export {}
+}